backend/requirements.txt
# =============================================================================
# Django Core
# =============================================================================
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
django-storages==1.14.2

# =============================================================================
# Authentication
# =============================================================================
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
dj-rest-auth==5.0.2

# =============================================================================
# Database
# =============================================================================
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# =============================================================================
# Caching & Task Queue
# =============================================================================
django-redis==5.4.0
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1

# =============================================================================
# Search
# =============================================================================
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5

# =============================================================================
# API & Serialization
# =============================================================================
drf-spectacular==0.27.0
drf-nested-routers==0.93.5
Pillow==10.1.0

# =============================================================================
# Production Server
# =============================================================================
gunicorn==21.2.0
whitenoise==6.6.0

# =============================================================================
# Payments
# =============================================================================
stripe==7.8.0

# =============================================================================
# Utilities
# =============================================================================
python-decouple==3.8
python-slugify==8.0.1
boto3==1.34.14
shortuuid==1.0.11

# =============================================================================
# Development & Testing
# =============================================================================
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.1
faker==22.0.0
flake8==7.0.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7
django-debug-toolbar==4.2.0
ipython==8.19.0

# =============================================================================
# Monitoring (Optional)
# =============================================================================
sentry-sdk==1.39.2
django-health-check==3.18.1
